+++ app/Preprocess.hs
@@ -3,8 +3,7 @@
 module Preprocess (main) where
 
 import Data.Tree (Tree(..), drawTree)
-import TreeBuilder (buildTreeSafe, buildParentChildMap, buildParentMap)
-import PreprocessUtils (loadEdges, distanceBetween)
+import TreeUtils (buildTreeSafe, buildParentChildMap, buildParentMap, loadEdges, distanceBetween)
 import qualified Data.Map.Strict as Map
 
 main :: IO ()
@@ -14,8 +13,8 @@ main = do
       pmap = buildParentMap pairs
       tree = buildTreeSafe cmap "entity"
 
-  putStrLn "ツリー構造をロードしました!"
-  putStrLn $ drawTree tree
+  -- putStrLn "Tree structure:"
+  -- putStrLn $ drawTree tree
 
   let nodeA = "チェッカー"
       nodeB = "アタッチメント"
